+++ src/servicios/ServicioRevisionDiaria.java
@@ -4,7 +4,7 @@ import containers.Container;
 
 public class ServicioRevisionDiaria implements Servicio {
 	
-	private int precioFijo; 
+	private double precioFijo; 
 	private int diasEnLaTerminal;  // TODAVIA FALTA CALCULAR LA CANTIDAD DE DIAS
 	
 	public ServicioRevisionDiaria(int precioFijo, int diasEnLaTerminal) {
